backend/core/engines/gpu_engine.py
"""
GPU-accelerated engine using cuQuantum (NVIDIA RTX).
Only used when cuquantum package is installed.
"""
import asyncio
from api.schemas.request import SimulateRequest
from api.schemas.response import SimulateResponse, ComplexNum

class GPUStateVectorEngine:
    """Uses cuquantum for GPU-accelerated state vector simulation"""

    def __init__(self):
        try:
            import cuquantum
            self.available = True
        except ImportError:
            self.available = False

    async def run(self, req: SimulateRequest) -> SimulateResponse:
        if not self.available:
            from core.engines.state_vector import StateVectorEngine
            return await StateVectorEngine().run(req)
        return await asyncio.to_thread(self._run_gpu_sync, req)

    def _run_gpu_sync(self, req: SimulateRequest) -> SimulateResponse:
        import cuquantum
        import numpy as np
        # cuQuantum circuit execution
        # (Full implementation requires cuquantum SDK installed with CUDA)
        raise NotImplementedError("cuQuantum implementation pending GPU environment setup")
