backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.2
pydantic-settings==2.5.2
numpy==1.26.4
qiskit==1.2.4
qiskit-aer==0.15.0
mqt.ddsim==1.22.0
python-dotenv==1.0.1
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
